requirements.txt
genanki
python-docx
pandas
flask
requests
